requirements.txt
Flask==2.1.3
Jinja2==3.1.1
Flask-Login==0.6.1
oauthlib==3.0.1
pyairtable==1.0.0.post1
requests==2.21.0
gunicorn==20.1.0
google-cloud-core==2.3.1
google-cloud-iot==2.6.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.2.0
google-cloud-functions==1.7.0
google-cloud-pubsub==2.13.0